ml-training/requirements.txt
# ML Training Environment Requirements
# Python 3.9+

# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.0.0

# Deep Learning (optional, for neural networks)
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2

# NLP (for sentiment analysis)
transformers==4.33.2
sentencepiece==0.1.99
datasets==2.14.4

# Experiment Tracking
mlflow==2.7.1
wandb==0.15.9

# Data Processing
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Database & API
psycopg2-binary==2.9.7
supabase==1.0.4
requests==2.31.0

# Utilities
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.25.2
tqdm==4.66.1
joblib==1.3.2

# Feature Engineering
feature-engine==1.6.2
imbalanced-learn==0.11.0

# Model Evaluation
shap==0.42.1
eli5==0.13.0


